cloud_functions/monthly_function/requirements.txt
google-cloud-datastore==1.9.0
requests==2.22.0
pydantic==1.6.2
